+++ src/app/api/auth/session/post.ts
@@ -69,5 +69,3 @@ export async function POST(request: NextRequest) {
     );
   }
 }
-
-export async function DELETE() {}
