+++ src/auth/components/common/password/PasswordSection.tsx
